requirements.txt
pytesseract==0.3.10
Pillow==9.4.0
pandas==1.5.3
openpyxl==3.1.2
